requirements.txt
# Core ML dependencies
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0,<4.30.0
numpy>=1.21.0,<2.0.0
#opencv-python>=4.5.0

# DICOM processing
pydicom>=2.3.0

# Image processing
matplotlib>=3.5.0
Pillow>=8.3.0

# API framework (for service mode)
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5

# Utilities
requests>=2.25.0
typing-extensions>=4.5.0
#safetensors>=0.3.0,<0.4.0

# Additional dependencies found in code
onnxruntime>=1.12.0  # Used in main.py